Update and rename .gitignore to SecretServer.java
Question #0 - A No-Brainer

Run the SecretServer and SecretClient.

Start the Server first with port number as a parameter, then run the Client with Server IP address (localhost) and Serer port number as parameters.

Observe that the length of the SECRET may vary, but the protocol ensures that the entire SECRET is sent to the Client.



Question #1 - Stop-and-Wait ARQ Normal Operation

Refer to the RDT General Protocol Requirements.

The SecretClient currently sends an extra "HELLO" to receive the "END" notification from the SecretServer, i.e. 9 HELLO's for 8 secret messages.

However, in the Protocol specification, the "END" notification is sent with the last secret message, i.e. there should be 8 HELLO's for 8 secret messages.

(a) Draw the flow diagram for the RDT Protocol, indicating the payload content.  Assume that the filename is HELLO.

(b) Modify SecretServer.java and SecretClient.java to follow the Protocol specification.